+++ Carlos/CacheProvider.swift
@@ -8,7 +8,7 @@ public class CacheProvider {
   - returns: An initialized and configured CacheLevel that takes NSURL keys and stores NSData values. Network requests are pooled for efficiency
   */
   public static func dataCache() -> BasicCache<NSURL, NSData> {
-    return MemoryCacheLevel() >>> DiskCacheLevel() >>> NetworkFetcher().pooled()
+    return MemoryCacheLevel() >>> (DiskCacheLevel() >>> NetworkFetcher()).pooled()
   }
   
   /**