+++ Carlos/NetworkFetcher.swift
@@ -13,80 +13,77 @@ public enum NetworkFetcherError: ErrorType {
 
 /// This class is a network cache level, mostly acting as a fetcher (meaning that calls to the set method won't have any effect). It internally uses NSURLSession to retrieve values from the internet
 public class NetworkFetcher: Fetcher {
+  private static let ValidStatusCodes = 200..<300
+  private let lock: ReadWriteLock = PThreadReadWriteLock()
+  
   /// The network cache accepts only NSURL keys
   public typealias KeyType = NSURL
   
   /// The network cache returns only NSData values
   public typealias OutputType = NSData
   
-  private class Request {
-    private static let ValidStatusCodes = 200..<300
-    private let URL : NSURL
-    private var task : NSURLSessionDataTask? = nil
-    
-    init(URL: NSURL, success succeed : (NSData, Request) -> (), failure fail : ((ErrorType, Request) -> ())) {
-      self.URL = URL
-      self.task = NSURLSession.sharedSession().dataTaskWithURL(URL) {[weak self] (data, response, error) in
-        if let strongSelf = self {
-          strongSelf.dataReceived(data, response: response, error: error, failure: fail, success: succeed)
-        }
-      }
-      task?.resume()
+  private func validate(response: NSHTTPURLResponse, withData data: NSData) -> Bool {
+    var responseIsValid = true
+    let expectedContentLength = response.expectedContentLength
+    if expectedContentLength > -1 {
+      responseIsValid = Int64(data.length) >= expectedContentLength
     }
+    return responseIsValid
+  }
+  
+  private func startRequest(URL: NSURL) -> Result<NSData> {
+    let result = Result<NSData>()
     
-    private func validate(response: NSHTTPURLResponse, withData data: NSData) -> Bool {
-      var responseIsValid = true
-      let expectedContentLength = response.expectedContentLength
-      if expectedContentLength > -1 {
-        responseIsValid = Int64(data.length) >= expectedContentLength
-      }
-      return responseIsValid
-    }
-    
-    private func dataReceived(data : NSData!, response : NSURLResponse!, error : NSError!, failure fail : ((ErrorType, Request) -> Void), success succeed : (NSData,Request) -> Void) {
+    let task = NSURLSession.sharedSession().dataTaskWithURL(URL) { [weak self] (data, response, error) in
+      guard let strongSelf = self else { return }
+      
       if let error = error {
         if error.domain != NSURLErrorDomain || error.code != NSURLErrorCancelled {
-          dispatch_async(dispatch_get_main_queue(), {
-            fail(error, self)
-          })
+          dispatch_async(dispatch_get_main_queue()) {
+            result.fail(error)
+          }
         }
       } else if let httpResponse = response as? NSHTTPURLResponse {
-        if !Request.ValidStatusCodes.contains(httpResponse.statusCode) {
+        if !NetworkFetcher.ValidStatusCodes.contains(httpResponse.statusCode) {
           dispatch_async(dispatch_get_main_queue()) {
-            fail(NetworkFetcherError.StatusCodeNotOk, self)
+            result.fail(NetworkFetcherError.StatusCodeNotOk)
           }
-        } else if !validate(httpResponse, withData: data) {
+        } else if let data = data where !strongSelf.validate(httpResponse, withData: data) {
           dispatch_async(dispatch_get_main_queue()) {
-            fail(NetworkFetcherError.InvalidNetworkResponse, self)
+            result.fail(NetworkFetcherError.InvalidNetworkResponse)
           }
         } else if let data = data {
           dispatch_async(dispatch_get_main_queue()) {
-            succeed(data,self)
+            result.succeed(data)
           }
         } else {
           dispatch_async(dispatch_get_main_queue()) {
-            fail(NetworkFetcherError.NoDataRetrieved, self)
+            result.fail(NetworkFetcherError.NoDataRetrieved)
           }
         }
       }
     }
+    
+    result.onCancel {
+      task.cancel()
+    }
+    
+    task.resume()
+    
+    return result
   }
 
-  private lazy var lockQueue: dispatch_queue_t = {
-    return dispatch_queue_create(CarlosGlobals.QueueNamePrefix + "networkfetcher", DISPATCH_QUEUE_SERIAL)
-  }()
+  private var pendingRequests: [Result<OutputType>] = []
 
-  private var pendingRequests: [Request] = []
-
-  private func addPendingRequest(request: Request) {
-    dispatch_async(lockQueue) {
+  private func addPendingRequest(request: Result<OutputType>) {
+    lock.withWriteLock {
       self.pendingRequests.append(request)
     }
   }
 
-  private func removePendingRequests(request: Request) {
-    dispatch_async(lockQueue) {
-      if let idx = self.pendingRequests.enumerate().filter({ $1 === request }).first?.index {
+  private func removePendingRequests(request: Result<OutputType>) {
+    if let idx = lock.withReadLock({ self.pendingRequests.indexOf({ $0 === request }) }) {
+      lock.withWriteLock {
         self.pendingRequests.removeAtIndex(idx)
       }
     }
@@ -105,18 +102,20 @@ public class NetworkFetcher: Fetcher {
   - returns: A Result that you can use to get the asynchronous results of the network fetch
   */
   public func get(key: KeyType) -> Result<OutputType> {
-    let result = Result<OutputType>()
+    let result = startRequest(key)
+      
+    result
+      .onSuccess { _ in
+        Logger.log("Fetched \(key) from the network fetcher")
+        self.removePendingRequests(result)
+      }
+      .onFailure { _ in
+        Logger.log("Failed fetching \(key) from the network fetcher", .Error)
+        self.removePendingRequests(result)
+      }
     
-    let request = Request(URL: key, success: { data,request in
-      Logger.log("Fetched \(key) from the network fetcher")
-      result.succeed(data)
-      self.removePendingRequests(request)
-    }, failure: { error, request in
-      Logger.log("Failed fetching \(key) from the network fetcher", .Error)
-      result.fail(error)
-      self.removePendingRequests(request)
-    })
-    self.addPendingRequest(request)
+    self.addPendingRequest(result)
+    
     return result
   }
 }